+++ blog/admin.py
@@ -1,3 +1,7 @@
 from django.contrib import admin
+from .models import User, Posts
 
 # Register your models here.
+
+admin.site.register(User)
+admin.site.register(Posts)
