+++ blog/templates/blog/home.html
@@ -61,7 +61,7 @@
                     <div class="card-body">
                         <h5 class="card-title">{{ post.title }}</h5>
                         <h6 class="card-subtitle mb-2 text-muted">By {{ post.author }} on {{ post.date_posted }}</h6>
-                        <p class="card-text">{{ post.content }}</p>
+                        <p class="card-text">{{ post.content | truncatechars:100 }}</p>
                         <a href="#" class="card-link">Read More</a>
                     </div>
                 </div>
