+++ blog/views.py
@@ -1,35 +1,7 @@
 from django.shortcuts import render
+from .models import Posts
 
-posts =  [
-
-    {
-        'author': 'Hilary Gee',
-        'title': 'Uncontrolled Tall Buildings',
-        'content': 'This is a blog post about the tall builds, which are not supposed to be around',
-        'date_posted': 'August 27, 2021',
-
-    }, 
-    {
-        'author': 'Jack M',
-        'title': 'Environmental Polution',
-        'content': 'This is a blog post about Environmental Polution',
-        'date_posted': 'August 27, 2023', 
-    },
-    {
-        'author': 'Loisa Saisa',
-        'title': 'Noise Pollution',
-        'content': 'This is a blog post about Noise Pollution',
-        'date_posted': 'August 27, 2023', 
-    },
-    {
-         
-        'author': 'Ahmed Nule',
-        'title': 'Poor Waste Management',
-        'content': 'This is a blog post about poor waste management',
-        'date_posted': 'August 27, 2023', 
-    
-    }
-]
+posts = Posts.objects.all()
 
 def home(request):
      context={
